move strategy out of main
