+++ lib/bombadil-cli/src/main.rs
@@ -2,14 +2,14 @@ mod duration;
 mod inspect_server;
 mod output_path;
 mod render;
+mod strategy;
 
 use ::url::Url;
-use anyhow::{Result, bail};
-use bombadil::specification::{convert::ToInternal, domain::Snapshot};
+use anyhow::Result;
+use bombadil::specification::convert::ToInternal;
 use clap::{Args, Parser};
 use serde_json as json;
 use std::{
-    collections::VecDeque,
     path::{Path, PathBuf},
     str::FromStr,
     time::{Duration, SystemTime},
@@ -21,15 +21,17 @@ use tokio::{fs::File, io::BufReader};
 use bombadil::{
     browser::{
         BrowserOptions, DebuggerOptions, Emulation, LaunchOptions,
-        actions::BrowserAction, state::BrowserState,
+        actions::BrowserAction,
     },
     instrumentation::InstrumentationConfig,
-    runner::{ControlFlow, RunStrategy, Runner},
-    specification::{convert::ToSchema, verifier::Specification},
+    runner::Runner,
+    specification::verifier::Specification,
     styled,
-    trace::{PropertyViolation, writer::TraceWriter},
+    trace::writer::TraceWriter,
 };
-use bombadil_schema::{markup, schema};
+use bombadil_schema::schema;
+
+use crate::strategy::{ExitReason, TestMode, TestResult, TestStrategy};
 
 /// Property-based testing for web UIs
 #[derive(Parser)]
@@ -398,7 +400,7 @@ async fn test(
 
     let deadline = shared_options.time_limit.map(|d| SystemTime::now() + d);
 
-    let mut strategy = MainStrategy {
+    let mut strategy = TestStrategy {
         mode,
         writer: TraceWriter::initialize(output_path.clone()).await?,
         exit_on_violation: shared_options.exit_on_violation,
@@ -470,250 +472,4 @@ async fn test(
     }
 
     Ok(())
-}
-
-enum TestMode {
-    RandomWalk,
-    Reproduce(VecDeque<BrowserAction>),
-}
-
-struct MainStrategy {
-    mode: TestMode,
-    writer: TraceWriter,
-    exit_on_violation: bool,
-    test_start: Option<bombadil_schema::Time>,
-    deadline: Option<SystemTime>,
-    output_path: PathBuf,
-    violations_count: u64,
-}
-
-#[derive(Clone, Copy, Debug)]
-enum ExitReason {
-    ExitOnViolation,
-    TimeLimit,
-    Interrupted,
-    Reproduced,
-}
-
-#[derive(Clone, Copy, Debug)]
-struct TestResult {
-    exit_reason: ExitReason,
-    violations_count: u64,
-}
-
-impl RunStrategy for MainStrategy {
-    type StopValue = TestResult;
-
-    async fn on_new_state(
-        &mut self,
-        state: &BrowserState,
-        last_action: Option<&BrowserAction>,
-        snapshots: &[Snapshot],
-        violations: &[PropertyViolation],
-    ) -> anyhow::Result<ControlFlow<Self::StopValue>> {
-        let test_start = *self.test_start.get_or_insert(
-            bombadil_schema::Time::from_system_time(state.timestamp),
-        );
-
-        if let Some(action) = last_action {
-            println!(
-                "{} {}",
-                render::format_timestamp(state.timestamp, test_start),
-                render::format_action(action)
-            );
-        }
-
-        self.violations_count += violations.len() as u64;
-        for violation in violations {
-            log::info!("violation of property `{}`", violation.name);
-            let api_violation = violation.to_schema();
-            let markup = markup::render_violation(&api_violation);
-            let text = styled::markup_to_styled(&markup, test_start);
-            println!(
-                "\n{}\n\n{}\n",
-                styled::maybe_red(styled::maybe_bold(format!(
-                    "{} was violated:",
-                    violation.name
-                ))),
-                text
-            );
-        }
-
-        self.writer
-            .write(state, last_action, snapshots, violations)
-            .await?;
-
-        if self.violations_count > 0 && self.exit_on_violation {
-            return Ok(ControlFlow::Stop(TestResult {
-                exit_reason: ExitReason::ExitOnViolation,
-                violations_count: self.violations_count,
-            }));
-        }
-
-        if let TestMode::Reproduce(browser_actions) = &self.mode
-            && browser_actions.is_empty()
-        {
-            return Ok(ControlFlow::Stop(TestResult {
-                exit_reason: ExitReason::Reproduced,
-                violations_count: self.violations_count,
-            }));
-        }
-
-        if let Some(deadline) = self.deadline
-            && state.timestamp >= deadline
-        {
-            log::info!("time limit reached, stopping");
-            return Ok(ControlFlow::Stop(TestResult {
-                exit_reason: ExitReason::TimeLimit,
-                violations_count: self.violations_count,
-            }));
-        }
-
-        Ok(ControlFlow::Continue)
-    }
-
-    async fn on_interrupted(&mut self) -> anyhow::Result<Self::StopValue> {
-        Ok(TestResult {
-            exit_reason: ExitReason::Interrupted,
-            violations_count: self.violations_count,
-        })
-    }
-
-    async fn pick_action(
-        &mut self,
-        tree: bombadil::tree::Tree<BrowserAction>,
-    ) -> Result<BrowserAction> {
-        match &mut self.mode {
-            TestMode::RandomWalk => Ok(tree.pick(&mut rand::rng())?.clone()),
-            TestMode::Reproduce(actions_original) => {
-                if let Some(action_original) = actions_original.pop_front() {
-                    let available_actions = tree.values();
-                    let action_reconciled = available_actions
-                        .iter()
-                        .filter_map(|action| {
-                            reconcile_reproducible_action(
-                                action,
-                                &action_original,
-                            )
-                        })
-                        .min_by(|(_, a), (_, b)| a.total_cmp(b))
-                        .map(|(action, _)| action);
-
-                    if let Some(action) = action_reconciled {
-                        Ok(action)
-                    } else {
-                        println!(
-                            "\n{}\n\n{}\n\n{}\n\n{}\n",
-                            styled::maybe_red(styled::maybe_bold(
-                                "no match for original:".into()
-                            )),
-                            render::format_action(&action_original),
-                            styled::maybe_red(styled::maybe_bold(
-                                "in the set of available actions:".into()
-                            )),
-                            tree.values()
-                                .iter()
-                                .map(render::format_action)
-                                .collect::<Vec<String>>()
-                                .join("\n")
-                        );
-                        bail!("reproduction and original test diverged!");
-                    }
-                } else {
-                    bail!(
-                        "no remaining actions in prefix to apply (this is a bug)"
-                    )
-                }
-            }
-        }
-    }
-}
-
-const RECONCILE_DISTANCE_MAX: f64 = 500.0;
-
-fn reconcile_reproducible_action(
-    candidate: &BrowserAction,
-    original: &BrowserAction,
-) -> Option<(BrowserAction, f64)> {
-    match (candidate, original) {
-        (BrowserAction::Back, BrowserAction::Back) => {
-            Some((BrowserAction::Back, 0.0))
-        }
-        (BrowserAction::Forward, BrowserAction::Forward) => {
-            Some((BrowserAction::Forward, 0.0))
-        }
-        (
-            BrowserAction::Click {
-                name: candidate_name,
-                content: candidate_content,
-                point: candidate_point,
-            },
-            BrowserAction::Click {
-                name: original_name,
-                content: original_content,
-                point: original_point,
-            },
-        ) if candidate_name == original_name
-            && candidate_content == original_content =>
-        {
-            let distance = (candidate_point.x - original_point.x)
-                .hypot(candidate_point.y - original_point.y);
-            (distance < RECONCILE_DISTANCE_MAX)
-                .then(|| (candidate.clone(), distance))
-        }
-        (
-            BrowserAction::DoubleClick {
-                name: candidate_name,
-                content: candidate_content,
-                point: candidate_point,
-                ..
-            },
-            BrowserAction::DoubleClick {
-                name: original_name,
-                content: original_content,
-                point: original_point,
-                ..
-            },
-        ) if candidate_name == original_name
-            && candidate_content == original_content =>
-        {
-            let distance = (candidate_point.x - original_point.x)
-                .hypot(candidate_point.y - original_point.y);
-            (distance < RECONCILE_DISTANCE_MAX)
-                .then(|| (candidate.clone(), distance))
-        }
-        (BrowserAction::TypeText { .. }, BrowserAction::TypeText { .. }) => {
-            Some((original.clone(), 0.0))
-        }
-        (BrowserAction::PressKey { .. }, BrowserAction::PressKey { .. }) => {
-            Some((original.clone(), 0.0))
-        }
-        (BrowserAction::ScrollUp { .. }, BrowserAction::ScrollUp { .. }) => {
-            Some((candidate.clone(), 0.0))
-        }
-        (
-            BrowserAction::ScrollDown { .. },
-            BrowserAction::ScrollDown { .. },
-        ) => Some((candidate.clone(), 0.0)),
-        (BrowserAction::Reload, BrowserAction::Reload) => {
-            Some((BrowserAction::Reload, 0.0))
-        }
-        (BrowserAction::Wait, BrowserAction::Wait) => {
-            Some((BrowserAction::Wait, 0.0))
-        }
-        (
-            BrowserAction::SetFileInputFiles {
-                selector: candidate_selector,
-                ..
-            },
-            BrowserAction::SetFileInputFiles {
-                selector: original_selector,
-                ..
-            },
-        ) if candidate_selector == original_selector => {
-            Some((original.clone(), 0.0))
-        }
-
-        _ => None,
-    }
 }
